mhealth-incognitus-backend/requirements.txt
flask~=2.3.3
minio~=7.1.16
openai-whisper
pandas
numpy
pymysql~=1.1.0
python-dateutil~=2.8.2